TkInter GUI support [WIP]
prototype for a basic TkInter GUI for the program. Right now it works only if there's already a credentials file and if the main password is 'test'. It still misses lots of functionality, right now it can only make the user visualize the list of credentials and copy to the clipboard the corresponding email or password by pressing the two buttons on the right of each list element.